add: error handling

+++ packages/api/src/router/like.ts
@@ -1,3 +1,5 @@
+import { Prisma } from "@prisma/client";
+import { TRPCError } from "@trpc/server";
 import { z } from "zod";
 
 import { createTRPCRouter, protectedProcedure } from "../trpc";
@@ -10,12 +12,36 @@ export const likeRouter = createTRPCRouter({
       }),
     )
     .mutation(async ({ input, ctx }) => {
-      await ctx.db.likes.create({
-        data: {
+      const existingLike = await ctx.db.likes.findFirst({
+        where: {
           postId: input.postId,
           userId: ctx.userId,
         },
       });
+      if (existingLike) {
+        throw new TRPCError({
+          code: "BAD_REQUEST",
+          message: "Post has already been liked by the user",
+        });
+      }
+      try {
+        await ctx.db.likes.create({
+          data: {
+            postId: input.postId,
+            userId: ctx.userId,
+          },
+        });
+      } catch (e) {
+        if (e instanceof Prisma.PrismaClientKnownRequestError) {
+          if (e.code === "P2025") {
+            throw new TRPCError({
+              code: "NOT_FOUND",
+              message: "Post not found",
+            });
+          }
+        }
+        throw e;
+      }
     }),
   unlikePost: protectedProcedure
     .input(
@@ -24,6 +50,18 @@ export const likeRouter = createTRPCRouter({
       }),
     )
     .mutation(async ({ input, ctx }) => {
+      const existingLike = await ctx.db.likes.findFirst({
+        where: {
+          postId: input.postId,
+          userId: ctx.userId,
+        },
+      });
+      if (!existingLike) {
+        throw new TRPCError({
+          code: "NOT_FOUND",
+          message: "Like not found",
+        });
+      }
       await ctx.db.likes.deleteMany({
         where: {
           postId: input.postId,
